WIP: Include property names in type mismatch exceptions.
Switch to validating the types of values prior to attempting to set the
properties at the object store level. This enables us to throw
exceptions that include more useful information, such as the name of the
property and the type it belongs to.

Does not yet attempt to work for JavaScriptCore, and has a few FIXMEs
for Node.